Icon optimization
A light bulb represents mostly ideas or some kind of that. This plugin
is about questions and a ranking of them. So I tried to crate an icon
that might make it a little bit more clear what the plugin is about. If
you like it, I would be happy if you want to use it.